requirements/base.txt
celery==4.4.2
#channels==2.4.0
#channels-redis==2.4.2
#Faker==4.0.3
flower==0.9.4
#jieba==0.42.1
PyYAML==5.3.1
redis==3.4.1
requests==2.23.0
#Whoosh==2.7.4

# Django
Django==3.0.5
#django-celery-beat==2.0.0
django-cors-headers==3.2.1
django-debug-toolbar==2.2
django-extensions==2.2.9
django-filter==2.2.0
#django-haystack==2.8.1
#django-mptt==0.11.0

# Django REST Framework
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
